added auto operate launcher command
